requirements.txt
python-dotenv
selenium
undetected-chromedriver